Split PaymentsTab into its own component
