+++ test/unit/about/preferencesTest.js
@@ -27,7 +27,7 @@ describe('Preferences component', function () {
     window.chrome = fakeElectron
     window.CustomEvent = {}
 
-    Preferences = require('../../../js/about/preferences')
+    Preferences = require('../../../js/about/preferences').AboutPreferences
   })
   after(function () {
     mockery.disable()
